refactor(web): move dbt-target-plugin.ts into artifact-source/
Plugin imports three files from artifact-source/ and serves as its
Vite entry point; moving it there makes the ownership explicit.

Update vite.config.ts, tsconfig.node.json, tsconfig.json, and
tsconfig.eslint.json to reflect the new path.
