+++ packages/dbt-tools/web/tsconfig.node.json
@@ -10,6 +10,6 @@
   "include": [
     "vite.config.ts",
     "vite.server.config.ts",
-    "src/dbt-target-plugin.ts"
+    "src/artifact-source/dbt-target-plugin.ts"
   ]
 }
